requirements.txt
numpy>=1.24.0
PyOpenGL>=3.1.7
pygame>=2.5.0
trimesh>=4.0.0
Pillow>=10.0.0
scipy>=1.11.0